+++ packages/react-components/react-components/src/unstable/index.ts
@@ -120,12 +120,21 @@ export {
   renderVirtualizer_unstable,
   useVirtualizerStyles_unstable,
   useIntersectionObserver,
+  useStaticVirtualizerMeasure,
+  VirtualizerScrollView,
+  virtualizerScrollViewClassNames,
+  useVirtualizerScrollView_unstable,
+  renderVirtualizerScrollView_unstable,
+  useVirtualizerScrollViewStyles_unstable,
 } from '@fluentui/react-virtualizer';
 export type {
   VirtualizerProps,
   VirtualizerState,
   VirtualizerSlots,
   VirtualizerChildRenderFunction,
+  VirtualizerScrollViewProps,
+  VirtualizerScrollViewState,
+  VirtualizerScrollViewSlots,
 } from '@fluentui/react-virtualizer';
 
 export {