+++ packages/react-components/react-virtualizer/src/hooks/useIntersectionObserver.ts
@@ -22,14 +22,6 @@ export const useIntersectionObserver = (
   setObserverInit: Dispatch<SetStateAction<IntersectionObserverInit | undefined>>;
   observer: MutableRefObject<IntersectionObserver | undefined>;
 } => {
-  // export const useIntersectionObserver = (
-  //   callback: IntersectionObserverCallback,
-  //   options?: IntersectionObserverInit,
-  // ): [
-  //   Dispatch<SetStateAction<Element[] | undefined>>,
-  //   Dispatch<SetStateAction<IntersectionObserverInit | undefined>>,
-  //   MutableRefObject<IntersectionObserver | undefined>,
-  // ] => {
   const observer = useRef<IntersectionObserver>();
   const [observerList, setObserverList] = useState<Element[]>();
   const [observerInit, setObserverInit] = useState<IntersectionObserverInit | undefined>(options);